@@ -1,10 +1,15 @@
 package com.qkopy.compressor
 
+import android.annotation.SuppressLint
 import android.app.Activity
 import android.content.Intent
+import android.graphics.Bitmap
 import android.graphics.BitmapFactory
+import android.media.ExifInterface
+import android.os.Build
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import android.provider.MediaStore
+import androidx.appcompat.app.AppCompatActivity
 import android.util.Log
 import android.view.View
 import com.qkopy.imagecompressor.Compressor
@@ -23,20 +28,29 @@ class MainActivity : AppCompatActivity() {
     private val PICK_IMAGE_REQUEST = 100
     private var actualImage: File? = null
     private var compressedImage: File? = null
+    private var compressedBitmap:Bitmap? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
         clearImage()
+
     }
 
     fun chooseImage(view: View) {
-        val intent = Intent(Intent.ACTION_GET_CONTENT)
+        val intent = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
+            Intent(Intent.ACTION_OPEN_DOCUMENT)
+        } else {
+            Intent(Intent.ACTION_GET_CONTENT)
+        }
+        intent.addCategory(Intent.CATEGORY_OPENABLE)
         intent.type = "image/*"
         startActivityForResult(intent, PICK_IMAGE_REQUEST)
+
     }
 
+    @SuppressLint("CheckResult")
     fun compressImage(view: View) {
         if (actualImage == null) {
             toast("Please choose an image!")
@@ -51,12 +65,18 @@ class MainActivity : AppCompatActivity() {
 
             // Compress image using RxJava in background thread
             Compressor(this@MainActivity)
+
+                //.setDestinationDirectoryPath(obbDir.path+File.separator+"image")
+                //.compressToBitmapAsFlowable(actualImage!!)
                 .compressToFileAsFlowable(actualImage!!)
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
                 .subscribe({ file ->
+
                     compressedImage = file
                     setCompressedImage()
+//                    compressedBitmap = file
+//                    setCompressedImageBitmap()
                 }, { throwable ->
                     throwable.printStackTrace()
                     toast("$throwable.message")
@@ -77,8 +97,18 @@ class MainActivity : AppCompatActivity() {
     private fun setCompressedImage() {
         compressed_image.setImageBitmap(BitmapFactory.decodeFile(compressedImage?.absolutePath))
         compressed_size.text = String.format("Size : %s", getReadableFileSize(compressedImage?.length()))
+        val exif = ExifInterface(compressedImage?.absolutePath)
+        compressed_meta.text = "${exif.getAttribute(ExifInterface.TAG_DATETIME)}\n${exif.getAttribute(ExifInterface.TAG_SOFTWARE)}\n" +
+                "${exif.getAttribute(ExifInterface.TAG_Y_RESOLUTION)}"
         //Toast.makeText(this, "Compressed image save in " + compressedImage?.path, Toast.LENGTH_LONG).show()
         Log.d("Compressor", "Compressed image save in " + compressedImage?.path)
+    }
+
+    private fun setCompressedImageBitmap() {
+        compressed_image.setImageBitmap(compressedBitmap)
+        compressed_size.text = String.format("Size : %s", getReadableFileSize(compressedBitmap?.byteCount?.toLong()))
+        //Toast.makeText(this, "Compressed image save in " + compressedImage?.path, Toast.LENGTH_LONG).show()
+        Log.d("Compressor", "Compressed image save in " + compressedBitmap?.config?.name)
     }
 
     private fun clearImage() {
@@ -98,6 +128,11 @@ class MainActivity : AppCompatActivity() {
                 if (actualImage!!.absolutePath != null) {
                     actual_image.setImageBitmap(BitmapFactory.decodeFile(actualImage!!.absolutePath))
                     actual_size.text = String.format("Size : %s", getReadableFileSize(actualImage!!.length()))
+                    val exif = ExifInterface(actualImage?.absolutePath)
+                    actual_meta.text = "${exif.getAttribute(ExifInterface.TAG_DATETIME)}\n${exif.getAttribute(ExifInterface.TAG_SOFTWARE)}\n" +
+                            "${exif.getAttribute(ExifInterface.TAG_Y_RESOLUTION)}"
+                    exif.toString()
+
                     clearImage()
                 }
 
